moved swing model out of common as only used in controlpanel
