+++ core/src/main/java/net/sourceforge/jnlp/controlpanel/NonEditableTableModel.java
@@ -31,7 +31,7 @@ library, you may extend this exception to your version of the library, but you
 are not obligated to do so.  If you do not wish to do so, delete this exception
 statement from your version.
 */
-package net.sourceforge.jnlp.util.ui;
+package net.sourceforge.jnlp.controlpanel;
 
 import javax.swing.table.DefaultTableModel;
 import java.util.Vector;
